pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["igh197 <ingyuhan497@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
ninja = "^1.11.1.1"


[tool.poetry.group.dev.dependencies]
black = "^24.3.0"
isort = "^5.13.2"
mypy = "^1.9.0"
django-stubs = "^4.2.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
